+++ src/components/Header/index.js
@@ -1,9 +1,9 @@
 import React, { useState } from 'react';
-import './Header.scss';
-import Logo from '../Logo/Logo';
-import CartBtn from '../Cart/Button/CartBtn';
-import CartContainer from '../Cart/Container/CartContainer';
-import Search from '../Search/Search';
+import './style.scss';
+import Logo from '../Logo';
+import CartBtn from '../Cart/Button';
+import CartContainer from '../Cart/Container';
+import Search from '../Search';
 
 const Header = () => {
   const [activeCart, setActiveCart] = useState(false);
