+++ src/components/Image/style.scss
