--- src/components/Logo/Logo.js
@@ -1,10 +1,0 @@
-import React from 'react';
-import './Logo.scss';
-
-const Logo = () => {
-  return (
-    <span className="Logo">Fashionista</span>
-  );
-}
-
-export default Logo;
