+++ src/components/Logo/index.js
@@ -1,0 +1,8 @@
+import React from 'react';
+import './style.scss';
+
+const Logo = () => {
+  return <span className='Logo'>Fashionista</span>;
+};
+
+export default Logo;
