+++ src/components/Logo/style.scss
