+++ src/components/ProductDescription/style.scss
