+++ src/index.js
@@ -1,9 +1,9 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import './index.scss';
-import Header from './components/Header/Header';
-import Home from './pages/Home/Home';
-import Footer from './components/Footer/Footer';
+import Header from './components/Header';
+import Home from './pages/Home';
+import Footer from './components/Footer';
 
 ReactDOM.render(
   <React.StrictMode>
